wip: resolve PR issues
