+++ src/theme.tsx
@@ -8,32 +8,6 @@ export const ColorModeContext = createContext({
     },
 });
 
-export const RTLContext = createContext({
-    toggleRTL: () => {},
-});
-
-export const useRTL = () => {
-    const [RTL, setRtl] = useState<"rtl" | "ltr">("ltr");
-
-    const RtlMode = useMemo(
-        () => ({
-            toggleRTL: () => {
-                document.dir = RTL === "ltr" ? "rtl" : "ltr";
-                return setRtl((prev) => (prev === "ltr" ? "rtl" : "ltr"));
-            },
-        }),
-        [RTL],
-    );
-
-    const RTLTheme = useMemo(
-        () => ({
-            direction: RTL,
-        }),
-        [RTL],
-    );
-    return [RTLTheme, RtlMode] as const;
-};
-
 export const useColorMode = () => {
     const prefersDarkMode = useMediaQuery("(prefers-color-scheme: dark)");
     const [mode, setMode] = useState<"light" | "dark">(
